requirements.txt
Flask==2.2.5
Flask-Cors==3.0.10
Flask-Limiter==3.3.1
requests==2.31.0
python-dotenv==1.0.1
redis==5.0.1
toml==0.10.2
google-generativeai==0.3.0

